refactor: add types util
