+++ react/lib/components/PaymentDialog/PaymentDialog.tsx
@@ -5,8 +5,7 @@ import { Theme, ThemeName, ThemeProvider, useTheme } from '../../themes';
 import Button, { ButtonProps } from '../Button/Button';
 import { WidgetContainer } from '../Widget/WidgetContainer';
 import { isValidCashAddress, isValidXecAddress } from '../../util/address';
-import { Transaction, currency } from '../../util/api-client';
-import { currencyObject } from '../../util/satoshis';
+import { currency, currencyObject, Transaction } from '../../util/types';
 
 export interface PaymentDialogProps extends ButtonProps {
   to: string;
